requirements/base.txt
# Django Core (Production Ready)
Django>=5.2.0,<5.3.0
djangorestframework>=3.14.0,<3.15.0
django-filter>=23.5,<24.0
django-cors-headers>=4.3.0,<5.0
django-crispy-forms>=2.0,<3.0
django-extensions>=3.2.0,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0

# Asynchronous Tasks and Caching
celery>=5.3.0,<6.0
redis>=5.0.0,<6.0
django-redis>=5.4.0,<6.0

# Production Server
gunicorn>=20.1.0,<22.0
whitenoise>=6.6.0,<7.0

# Database (PostgreSQL Ready)
psycopg[binary]>=3.2.0,<4.0
django-redis-sessions>=0.6.0,<1.0

# File Processing and Storage
Pillow>=10.0.0,<11.0
django-storages[boto3]>=1.14.0,<2.0
boto3>=1.34.0,<2.0
python-magic>=0.4.27,<1.0

# Configuration and Environment
python-dotenv>=1.0.0,<2.0
django-environ>=0.11.0,<1.0

# Data Processing and Reporting
pandas>=2.0.0,<3.0
openpyxl>=3.1.0,<4.0
xlsxwriter>=3.1.0,<4.0
reportlab>=4.0.0,<5.0
PyYAML>=6.0,<7.0

# AI and Machine Learning (Production)
google-generativeai>=0.3.0,<1.0
google-cloud-aiplatform>=1.38.0,<2.0
sentence-transformers>=2.2.0,<3.0
torch>=2.0.0,<3.0
numpy>=1.24.0,<2.0
faiss-cpu>=1.7.4,<2.0

# Security and Authentication
django-auditlog>=3.0.0,<4.0
django-axes>=6.1.0,<7.0
django-ratelimit>=4.1.0,<5.0
django-oauth-toolkit>=2.3.0,<3.0

# Database Structure
django-mptt>=0.16.0,<1.0

# Email Services (Production)
django-anymail>=10.0,<11.0

# Performance Monitoring
django-debug-toolbar>=4.2.0,<5.0  # Development only
silk>=5.0.0,<6.0  # Optional production monitoring

# Health and Monitoring
django-health-check>=3.17.0,<4.0

# Development Tools (for container builds)
bandit>=1.7.4,<2.0
safety>=2.3.0,<3.0
pip-audit>=2.5.0,<3.0
